add ability to 'force even' when merging pdfs
setting the force even flag to true when sending a merge message to gedoc will mean that gedoc will count the pages of each pdf before merging. If the page count is odd it will append a 'this page left intentionally blank' page to the end of the document. Every individual pdf will be an even number of pages before merging. This is important for double sided printing.

This commit also seems some dependency updates to solve some issues that go was having with importing dependencies of jaeger