store latexconfig in byte string and use TempDir for build folder
refactored all functions to manage the creation of their own build folder internally

avoids the need for the program to depend on a folder existing is a specific location. instead each function will create a tmpDir in /tmp and the latex function will insert the latex config from a byte array rather than having to read it from a file

The actual motivation for this change is that .latexmkrc wasn't being copied into the docker container. rather than debug docker the program was adjusted to be more robust